Completed refactoring of 'Frames' functionality!
