+++ frames/__init__.py
@@ -1,0 +1,10 @@
+from . import model
+from . import observer
+from ._frametools import make_frame
+from ._frametools import load_frame
+
+# def make_model(system, initials, **kwargs):
+#     builts = {'system': system}
+#     builts.update(initials)
+#     newModel = make_frame('model', builts, **kwargs)
+#     return newModel
