+++ frames/_frame.py
@@ -1,133 +1,4 @@
-import underworld as uw
-import tarfile
-import os
-import shutil
-import json
-import copy
-import glob
-
-from . import paths
-from . import utilities
-from . import disk
-from .wordhash import wordhash as wordhashFn
-from . import checkpoint
-from .utilities import message
-from .utilities import check_reqs
-from .visualisation import QuickFig
-
-def expose_tar(path):
-
-    tarpath = path + '.tar.gz'
-
-    print(path)
-
-    if uw.mpi.rank == 0:
-        assert os.path.isdir(path) or os.path.isfile(tarpath), \
-            "No model found at that directory!"
-
-    if uw.mpi.rank == 0:
-        if os.path.isfile(tarpath):
-            assert not os.path.isdir(path), \
-                "Conflicting archive and directory found."
-            message("Tar found - unarchiving...")
-            with tarfile.open(tarpath) as tar:
-                tar.extractall(path)
-            message("Unarchived.")
-            assert os.path.isdir(path), \
-                "Archive contained the wrong model file somehow."
-            os.remove(tarpath)
-
-def load_json(jsonName, path):
-    filename = jsonName + '.json'
-    jsonDict = {}
-    if uw.mpi.rank == 0:
-        with open(os.path.join(path, filename)) as json_file:
-            jsonDict = json.load(json_file)
-    jsonDict = uw.mpi.comm.bcast(jsonDict, root = 0)
-    return jsonDict
-
-load_inputs = load_json
-
-def load_jsons(jsonNames, path):
-    jsons = {}
-    if uw.mpi.rank == 0:
-        for jsonName in jsonNames:
-            jsons[jsonName] = load_json(
-                jsonName,
-                path
-                )
-    jsons = uw.mpi.comm.bcast(jsons, root = 0)
-    return jsons
-
-def load_builts(path):
-    builtsDir = os.path.join(path, 'builts')
-    names = set()
-    if uw.mpi.rank == 0:
-        assert os.path.isdir(builtsDir)
-        files = os.listdir(builtsDir)
-        for file in os.listdir(builtsDir):
-            if file.endswith('.json'):
-                builtName = os.path.splitext(
-                    os.path.basename(file)
-                    )[0]
-                names.add(builtName)
-    names = uw.mpi.comm.bcast(names, root = 0)
-    builts = {}
-    for name in sorted(names):
-        inputDict = load_json(name, builtsDir)
-        scriptModules = []
-        index = 0
-        while True:
-            scriptName = name + '_' + str(index) + '.py'
-            scriptPath = os.path.join(
-                builtsDir,
-                scriptName
-                )
-            fileCheck = False
-            if uw.mpi.rank == 0:
-                fileCheck = os.path.isfile(scriptPath)
-            fileCheck = uw.mpi.comm.bcast(fileCheck, root = 0)
-            if fileCheck:
-                scriptModule = utilities.local_import(
-                    scriptPath
-                    )
-                scriptModules.append(scriptModule)
-                index += 1
-            else:
-                break
-        built = scriptModule.build(
-            *scriptModules[1:],
-            **inputDict
-            )
-        builts[name] = built
-    return builts
-
-def make_stamps(builts):
-    scripts = {}
-    inputs = {}
-    for builtName, built in sorted(builts.items()):
-        scripts[builtName] = built.scripts
-        inputs[builtName] = built.inputs
-    assert inputs.keys() == scripts.keys()
-    stamps = {}
-    if uw.mpi.rank == 0:
-        toHash = {}
-        for key in sorted(inputs.keys()):
-            inputList = inputs[key]
-            scriptList = scripts[key]
-            toHash[key + 'inputs'] = inputList
-            toHash[key + 'scripts'] = [
-                open(script) for script in scriptList
-                ]
-        stamps = {
-            key: utilities.hashstamp(val) \
-                for key, val in sorted(toHash.items())
-            }
-        stamps['all'] = utilities.hashstamp(stamps)
-        for stampKey, stampVal in stamps.items():
-            stamps[stampKey] = [stampVal, wordhashFn(stampVal)]
-    stamps = uw.mpi.comm.bcast(stamps, root = 0)
-    return stamps
+from ._frametools import *
 
 class Frame:
 
@@ -143,6 +14,7 @@ class Frame:
         'update',
         'initialise',
         'builts',
+        'info'
         }
 
     _autobackup = True
@@ -175,6 +47,7 @@ class Frame:
             figs = self.figs,
             dataCollectors = self.collectors,
             builts = self.builts,
+            info = self.info,
             inFrames = self.inFrames,
             )
