+++ frames/observer.py
