+++ tests/framestest.py
@@ -1,5 +1,5 @@
 from .. import initials
-from .. import model
+from .. import frames
 from .. import systems
 from ..utilities import message
 import underworld as uw
@@ -9,19 +9,22 @@ from ..paths import TestDir
 
 def testfn():
     with TestDir() as outputPath:
-        model0 = model.make_model(
+        model0 = frames.frame.make_frame(
+            'model',
             systems.arrhenius.build(res = 32, f = 0.5),
             {'temperatureField': initials.sinusoidal.build()},
             outputPath = outputPath
             )
-        model1 = model.make_model(
+        model1 = frames.frame.make_frame(
+            'model',
             systems.arrhenius.build(res = 32, f = 1.),
             {'temperatureField': initials.sinusoidal.build()},
             # {'temperatureField': initials.load.build(model0, 'temperatureField')},
             outputPath = outputPath
             )
         model1.checkpoint()
-        model2 = model.make_model(
+        model2 = frames.frame.make_frame(
+            'model',
             systems.arrhenius.build(res = 32, f = 1.),
             {'temperatureField': initials.sinusoidal.build()},
             # {'temperatureField': initials.load.build(model0, 'temperatureField')},
@@ -31,5 +34,5 @@ def testfn():
         model2.unarchive()
         model2.archive()
         model2.checkpoint()
-        model3 = model.load_model(outputPath, model1.instanceID)
+        model3 = frames.frame.load_frame(outputPath, model1.instanceID)
         message('Success!')
